requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
requests==2.28.1
mysqlclient==2.1.1
python-dotenv==0.21.0
python-decouple==3.6
cryptography==38.0.3
PyJWT==2.6.0